+++ moduleframework/helpfile_linter.py
@@ -1,6 +1,5 @@
 from __future__ import print_function
 
-import mistune
 import os
 from moduleframework import common
 
@@ -27,40 +26,25 @@ class HelpMDLinter(object):
     def __init__(self, dir_name="../"):
         help_md_file = get_help_md_file(dir_name)
         if help_md_file:
-            renderer = mistune.Renderer(escape=False,
-                                        hard_wrap=False,
-                                        parse_block_html=False,
-                                        parse_inline_html=False)
-            md = mistune.Markdown(renderer=renderer)
             with open(help_md_file, 'r') as f:
-                self.help_md = md.parse(f.read())
+                lines = f.readlines()
+                # Count with all lines which begins with #
+                self.help_md = [x.strip() for x in lines if x.startswith('#')]
+                # Count with all lines which begins with %
+                self.help_md.extend([x.strip() for x in lines if x.startswith('%')])
         else:
             self.help_md = None
-        print(self.help_md)
 
-    def get_image_name(self):
-        pass
+    def get_image_maintainer_name(self, name):
+        name = '% %s' % name
+        if name.upper() in self.help_md:
+            return True
+        else:
+            return False
 
-    def get_maintainer(self):
-        pass
-
-    def get_date(self):
-        pass
-
-    def get_name(self):
-        pass
-
-    def get_description(self):
-        pass
-
-    def get_usage(self):
-        pass
-
-    def get_environment_variables(self):
-        pass
-
-    def get_labels(self):
-        pass
-
-    def get_security_implications(self):
-        pass
+    def get_tag(self, name):
+        name = '# %s' % name
+        if name.upper() in self.help_md:
+            return True
+        else:
+            return False
